Move field validation code to `/validateField`
