+++ packages/validation/src/validateField/restrictions/scriptValidation.ts
