+++ packages/validation/src/validateField/restrictions/testCodeList.ts
@@ -17,29 +17,34 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { RestrictionRange } from 'dictionary';
+import type { RestrictionCodeList } from 'dictionary';
 import { invalid, valid } from '../../types/testResult';
-import { isWithinRange } from '../../utils/isWithinRange';
-import { rangeToText } from '../../utils/rangeToText';
-import type { FieldRestrictionSingleValueTest, FieldRestrictionTest } from '../FieldRestrictionTest';
+import type { FieldRestrictionSingleValueTestFunction, FieldRestrictionTestFunction } from '../FieldRestrictionTest';
 import { createFieldRestrictionTestForArrays } from './createFieldRestrictionTestForArrays';
 
-const testRangeSingleValue: FieldRestrictionSingleValueTest<RestrictionRange> = (rule, value) => {
-	if (typeof value !== 'number') {
-		// only apply range tests to numbers
+const testCodeListSingleValue: FieldRestrictionSingleValueTestFunction<RestrictionCodeList> = (rule, value) => {
+	// TODO: this can be sped up by using a set for the rule instead of an array. finding an element in the set is faster.
+	if (!(typeof value === 'string' || typeof value === 'number')) {
+		// only apply code list check to strings and numbers
 		return valid();
 	}
 
-	if (isWithinRange(rule, value)) {
-		return valid();
+	// We want to compare strings after removing whitespace and converting both the option and the value to lowercase
+	const testValue = typeof value === 'string' ? value.trim().toLowerCase() : value;
+
+	for (const option of rule) {
+		const testOption = typeof option === 'string' ? option.trim().toLowerCase() : option;
+		if (testOption === testValue) {
+			return valid();
+		}
 	}
-	return invalid({ message: `The value must be within the range: ${rangeToText(rule)}` });
+	return invalid({ message: `The value for this field must match an option from the list.` });
 };
 
-const testRangeArray = createFieldRestrictionTestForArrays(
-	testRangeSingleValue,
-	(rule) => `All values in the array must be within the range: ${rangeToText(rule)}`,
+const testCodeListArray = createFieldRestrictionTestForArrays(
+	testCodeListSingleValue,
+	`All values in this field must match an option from the list.`,
 );
 
-export const testRange: FieldRestrictionTest<RestrictionRange> = (rule, value) =>
-	Array.isArray(value) ? testRangeArray(rule, value) : testRangeSingleValue(rule, value);
+export const testCodeList: FieldRestrictionTestFunction<RestrictionCodeList> = (rule, value) =>
+	Array.isArray(value) ? testCodeListArray(rule, value) : testCodeListSingleValue(rule, value);
